+++ omniload/target/filesystem/api.py
@@ -1,7 +1,12 @@
 from omniload.target.filesystem.local import LocalFilesystemDestination
-from omniload.target.filesystem.remote import GCSDestination, S3Destination
+from omniload.target.filesystem.remote import (
+    AzureDestination,
+    GCSDestination,
+    S3Destination,
+)
 
 __all__ = [
+    "AzureDestination",
     "GCSDestination",
     "LocalFilesystemDestination",
     "S3Destination",